moved cxf sample tests to their own file
