+++ crates/bitwarden-exporters/src/cxf/mod.rs
@@ -16,6 +16,8 @@ mod api_key;
 mod card;
 mod editable_field;
 mod identity;
+#[cfg(test)]
+mod import_sample_tests;
 mod login;
 mod note;
 mod wifi;
